# docs/api/project-management-openapi.yaml
openapi: 3.0.3
info:
  title: VideoPlanet Project Management API
  description: |
    프로젝트 생성, 일정 관리, 템플릿 시스템을 위한 종합적인 API
    
    ## 주요 기능
    - 프로젝트 생성 및 관리 (RBAC 권한 시스템)
    - 프로젝트 일정 관리 (마일스톤, 데드라인, 리소스 할당)
    - 프로젝트 템플릿 시스템 (산업별 템플릿, 커스터마이제이션)
    - 충돌 감지 및 해결
    - 실시간 협업 및 알림
    
    ## 인증
    모든 API는 Bearer JWT 토큰 인증이 필요합니다.
    
    ## 성능 요구사항
    - 프로젝트 생성: 2초 이내
    - 일정 조회: 500ms 이내  
    - 권한 체크: 100ms 이내
    - 동시 접근: 100명 이상 지원
    
  version: 1.0.0
  contact:
    name: VideoPlanet API Team
    url: https://api.vlanet.net/docs
    email: api@vlanet.net
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.vlanet.net
    description: Production server
  - url: https://staging-api.vlanet.net  
    description: Staging server
  - url: http://localhost:8000
    description: Development server
security:
  - bearerAuth: []
paths:
...
  /api/projects:
    get:
      summary: 프로젝트 목록 조회
      description: |
        사용자가 접근 권한을 가진 프로젝트 목록을 조회합니다.
        필터링, 정렬, 페이지네이션을 지원합니다.
      tags:
        - Projects
      parameters:
        - name: page
          in: query
          description: 페이지 번호 (1부터 시작)
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: 페이지당 항목 수
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: search
          in: query
          description: 프로젝트 이름/설명 검색
          schema:
            type: string
            maxLength: 100
        - name: status
          in: query
          description: 프로젝트 상태 필터 (다중 선택 가능)
          schema:
            type: array
            items:
              $ref: '#/components/schemas/ProjectStatus'
        - name: category
          in: query
          description: 프로젝트 카테고리 필터 (다중 선택 가능)
          schema:
            type: array
            items:
              $ref: '#/components/schemas/ProjectCategory'
        - name: sortBy
          in: query
          description: 정렬 기준
          schema:
            type: string
            enum: [name, createdAt, updatedAt, status]
            default: updatedAt
        - name: sortOrder
          in: query
          description: 정렬 순서
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: 프로젝트 목록 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: 새 프로젝트 생성
      description: |
        새 프로젝트를 생성합니다. 생성자는 자동으로 Owner 권한을 받습니다.
        프로젝트 생성 시 기본 폴더 구조(/media/projects/{project_id}/)가 자동으로 생성됩니다.
      tags:
        - Projects
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
            example:
              name: "마케팅 비디오 프로젝트"
              description: "Q1 제품 마케팅을 위한 브랜드 비디오 제작"
              category: "marketing"
              tags: ["marketing", "brand", "video"]
              settings:
                visibility: "private"
                allowComments: true
                allowDownloads: false
      responses:
        '201':
          description: 프로젝트 생성 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  project:
                    $ref: '#/components/schemas/Project'
                  folderPath:
                    type: string
                    description: 생성된 프로젝트 폴더 경로
                    example: "/media/projects/proj_1234567890_abc123"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ProjectStatus:
      type: string
      enum:
        - draft
        - planning
        - in_progress
        - review
        - completed
        - cancelled
        - on_hold
    ProjectMemberRole:
      type: string
      enum:
        - owner
        - admin  
        - editor
        - reviewer
        - viewer
    ProjectCategory:
      type: string
      enum:
        - marketing
        - education
        - entertainment
        - corporate
        - documentary
        - other
...
    ProjectPermissions:
      type: object
      required:
        - canEdit
        - canDelete
        - canInviteMembers
        - canManageSettings
        - canUploadVideos
        - canViewAnalytics
      properties:
        canEdit:
          type: boolean
        canDelete:
          type: boolean
        canInviteMembers:
          type: boolean
        canManageSettings:
          type: boolean
        canUploadVideos:
          type: boolean
        canViewAnalytics:
          type: boolean
...
    ProjectMember:
      type: object
      required:
        - userId
        - role
        - joinedAt
        - permissions
      properties:
        userId:
          type: string
          format: uuid
        role:
          $ref: '#/components/schemas/ProjectMemberRole'
        joinedAt:
          type: string
          format: date-time
        permissions:
          $ref: '#/components/schemas/ProjectPermissions'
    Project:
      type: object
      required:
        - id
        - name
        - status
        - owner
        - members
        - settings
        - metadata
        - createdAt
        - updatedAt
        - isArchived
      properties:
        id:
          type: string
          format: uuid
          description: "UUID 기반 프로젝트 식별자"
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 1000
        status:
          $ref: '#/components/schemas/ProjectStatus'
        owner:
          $ref: '#/components/schemas/ProjectMember'
        members:
          type: array
          items:
            $ref: '#/components/schemas/ProjectMember'
        settings:
          type: object
          properties:
            visibility:
              type: string
              enum: [public, private, team]
            allowComments:
              type: boolean
            allowDownloads:
              type: boolean
        metadata:
          type: object
          properties:
            tags:
              type: array
              items:
                type: string
            category:
              $ref: '#/components/schemas/ProjectCategory'
            deliverables:
              type: array
              items:
                type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        isArchived:
          type: boolean
...
    CreateProjectRequest:
      type: object
      required:
        - name
        - category
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: "프로젝트 이름"
        description:
          type: string
          maxLength: 1000
          description: "프로젝트 설명"
        category:
          $ref: '#/components/schemas/ProjectCategory'
        tags:
          type: array
          items:
            type: string
          description: "프로젝트 태그"
        settings:
          type: object
          description: "프로젝트 설정"
...
    ProjectListResponse:
      type: object
      required:
        - projects
        - total
        - page
        - limit
        - hasNext
      properties:
        projects:
          type: array
          items:
            $ref: '#/components/schemas/Project'
        total:
          type: integer
          minimum: 0
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
        hasNext:
          type: boolean
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
        - timestamp
        - path
      properties:
        error:
          type: string
          description: "에러 코드"
        message:
          type: string
          description: "에러 메시지"
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
              value:
                type: string
        timestamp:
          type: string
          format: date-time
        path:
          type: string
          description: "API 경로"
  responses:
    BadRequest:
      description: 잘못된 요청
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "VALIDATION_ERROR"
            message: "요청 데이터가 올바르지 않습니다"
            details:
              - field: "name"
                message: "프로젝트 이름은 필수입니다"
                value: ""
            timestamp: "2025-01-15T10:30:00Z"
            path: "/api/projects"
    Unauthorized:
      description: 인증 필요
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "UNAUTHORIZED"
            message: "인증이 필요합니다"
            timestamp: "2025-01-15T10:30:00Z"
            path: "/api/projects"
...
    InternalServerError:
      description: 서버 내부 오류
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "INTERNAL_SERVER_ERROR"
            message: "서버에서 오류가 발생했습니다"
            timestamp: "2025-01-15T10:30:00Z"
            path: "/api/projects"
...